{fix} rewrite slashdash logic
Instead of allowing slashdash entries and children anywhere in a
node definition, we now handle slashdash as a modifier on the
entires and children. This more closely matches the specs.

As a result, it is now no longer possible to have slashdash
children and regular children, and slashdash children are no
longer valid when followed by (slashdash) entries.

Fixes #3