{feature} try to retain the original in serialization
Try to retain the original order of entries when serializing
an element that came from deserialization.
This only applies to entries, not children, because the order
of children might matter.

The original order of properties is retained completely.
The order of arguments cannot be, for the simple reason that
the order of arguments is likely to be significant.
The interleaving of properties and arguments is retained.